requirements.txt
opencv-python
numpy
pytesseract
google.generativeai
rapidfuzz
gradio